Replace deno-db with deno-postgres and raw sql queries
(Fine, I'll do it myself...)

This commit represents a branch to replace the denodb orm with deno-postgres and
a small db provider to access it.

Current DenoDB project had many issues and specially had some troubles sync the db
without dropping the info: https://github.com/eveningkid/denodb/issues/283, https://github.com/eveningkid/denodb/issues/258

So everytime the server restarted, the db was wiped out and therefore it was unusable for production.

After this change, the db is now "persistent" (and the code is arguably cleaner/straightforward)