Improve atractiveness of event descriptions
To make the descriptions more appealing/simpler we take only the first paragraph of every one. thus making every event more misterious and easier to read/scan.
Additionally, with event with long descriptions it is very clear that the context is scraped and the first paragraph of the events are usually nice and interesting enough.